requirements.txt
autopep8==1.4.3
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mecab-python3==0.996.1
numpy==1.15.4
pandas==0.23.4
pycodestyle==2.4.0
pylint==2.2.2
python-dateutil==2.7.5
pytz==2018.7
six==1.12.0
typed-ast==1.1.1
wrapt==1.10.11
